requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPdf2
chromadb
faiss-cpu
langchain_google_genai
